Run dmesgd when the device starts to collect boot-time reports
Normally we rely on BootReceiver watching the tracing pipe for KASAN/KFENCE reports and running dmesgd. But if an error is reported before system_server starts, nobody is waiting for it.
To handle this, we set the dmesgd.start property after the device finishes booting, so that dmesgd fires once to collect the early reports.

Bug: 254725406
Test: injected kernel bugs reported at boot time are collected by dmesgd
Change-Id: Iedb66705c77e0d4a08e1bc53fd1dc920aecc4035